NonLeaf: Insert
Added Insert for Nonlead. Used a lot of your code to implement insert,
with the 1 difference that I used memmove, which removes the need for a
temporary buffer. Also, since all array manipulations are going to be
done in units of sizeof(Id) + sizeof(int), I defined that as a constant
as well.